refactor: named identifier better
